requirements.txt
streamlit>=1.32.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-huggingface>=0.0.3
PyPDF2>=3.0.1
chromadb>=0.4.24
sentence-transformers>=2.6.1
python-docx>=1.1.0
edge-tts>=6.1.9
streamlit-mic-recorder>=0.0.3
SpeechRecognition>=3.10.0
pysqlite3-binary>=0.5.2.post3
